+++ wars/api-backoffice/src/main/java/it/govpay/backoffice/v1/beans/ListaStatisticheQuadrature.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaStatisticheQuadrature extends Lista<StatisticaQuadratura> {
 	
-	public ListaStatisticheQuadrature(List<StatisticaQuadratura> pagamentiPortale, URI requestUri, long count, long pagina, long limit) {
+	public ListaStatisticheQuadrature(List<StatisticaQuadratura> pagamentiPortale, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(pagamentiPortale, requestUri, count, pagina, limit);
 	}
