+++ wars/api-backoffice/src/main/java/it/govpay/backoffice/v1/beans/ListaUnitaOperative.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaUnitaOperative extends Lista<UnitaOperativa> {
 	
-	public ListaUnitaOperative(List<UnitaOperativa> pagamentiPortale, URI requestUri, long count, long pagina, long limit) {
+	public ListaUnitaOperative(List<UnitaOperativa> pagamentiPortale, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(pagamentiPortale, requestUri, count, pagina, limit);
 	}
