+++ wars/api-pagamento/src/main/java/it/govpay/pagamento/v1/beans/ListaDominiIndex.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaDominiIndex extends Lista<DominioIndex> {
 	
-	public ListaDominiIndex(List<DominioIndex> domini, URI requestUri, long count, long pagina, long limit) {
+	public ListaDominiIndex(List<DominioIndex> domini, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(domini, requestUri, count, pagina, limit);
 	}
